A4A MSD: address overview tier card review feedback
- Keep cache persistence for the agency membership query; only the
  active-agency query opts out, since tier, influenced revenue, and
  approval status are what must be fresh after a reload.
- Hoist the pending and rejected application cards out of
  TierOverviewCard into application-status-cards.tsx and let the
  overview body pick the state, slimming the tier card down to the
  approved state.
- Use the Heading primitive for the application cards' headings,
  matching OverviewCard.
- Give the VIP tier its own description instead of repeating Pro's.
